backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Moritz Scheck"] # Poetry requires an author
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = { version = ">=0.114.2,<1.0.0", extras = ["standard"] }
python-multipart = ">=0.0.7,<1.0.0"
tenacity = ">=8.2.3,<9.0.0"
pydantic = ">2.0"
alembic = ">=1.12.1,<2.0.0"
httpx = ">=0.25.1,<1.0.0"
psycopg = { version = ">=3.1.13,<4.0.0", extras = ["binary"] }
sqlmodel = ">=0.0.21,<1.0.0"
# Pin bcrypt until passlib supports the latest
bcrypt = "==4.0.1"
pydantic-settings = ">=2.2.1,<3.0.0"
sentry-sdk = { version = ">=1.40.6,<2.0.0", extras = ["fastapi"] }
pyjwt = ">=2.8.0,<3.0.0"
google-genai = "^1.31.0"
pytz = "^2025.2"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.3,<8.0.0"
mypy = ">=1.8.0,<2.0.0"
ruff = ">=0.2.2,<1.0.0"
pre-commit = ">=3.6.2,<4.0.0"
types-passlib = ">=1.7.7.20240106,<2.0.0.0"
coverage = ">=7.4.3,<8.0.0"

# --- The sections below are copied directly ---
pytest-cov = "^6.2.1"
hypothesis = "^6.138.13"

[build-system]
# This section is updated to use Poetry as the build backend
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]

ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true